// apps/legacy-hono-api/src/v1/schema.ts
const paymentProviderConfigurationProductSchema = z
  .union([
    ...paymentProviders.map((p) =>
      z.object({
        // providerId: z.literal(p.getId()),
        paymentProviderConfigurationId: z.string(),
        configuration: p.getProductConfigurationSchema()
      })
    )
  ])
  .meta({
    ref: 'PaymentProviderConfigurationProduct'
  });

export const attachProviderProductParamsSchema = z.object({
  productId: z.string()
});

export const attachProviderProductBodySchema =
  paymentProviderConfigurationProductSchema.meta({
    ref: 'AttachProviderProductBody'
  });

export const providerProductResponseSchema = z
  .object({
    providerProductKey: z.string(),
    providerConfiguration: paymentProviderConfigurationProductSchema
  })
  .meta({
    ref: 'ProviderProduct'
  });

export const getProviderProductsParamsSchema = z.object({
  productId: z.string()
});

export const updateProviderProductParamsSchema = z.object({
  paymentProviderConfigurationProductId: z.string()
});

export const updateProviderProductBodySchema =
  paymentProviderConfigurationProductSchema.meta({
    ref: 'UpdateProviderProductBody'
  });